dashboard/requirements.txt
Flask==2.3.3
gspread==5.11.3
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.95.0
python-dotenv==1.0.0
requests==2.31.0
